Move the general code to optimagic.typing.
