+++ src/optimagic/optimization/algorithm.py
@@ -1,9 +1,8 @@
-import functools
 import typing
 import warnings
 from abc import ABC, ABCMeta, abstractmethod
 from dataclasses import dataclass, replace
-from typing import Any, Callable, TypeVar
+from typing import Any
 
 import numpy as np
 import pydantic
@@ -16,67 +15,11 @@ from optimagic.optimization.history import History
 from optimagic.optimization.internal_optimization_problem import (
     InternalOptimizationProblem,
 )
-from optimagic.typing import AggregationLevel
-
-DataclassT = TypeVar("DataclassT")
-
-OPTION_VALIDATION_CONFIG = pydantic.ConfigDict(
-    arbitrary_types_allowed=True,
-    extra="forbid",
-    validate_default=True,
+from optimagic.typing import (
+    STRICT_VALIDATION_CONFIG,
+    AggregationLevel,
+    validated_dataclass,
 )
-"""Pydantic config for user-facing options: coerce generous inputs to strict types."""
-
-STRICT_VALIDATION_CONFIG = pydantic.ConfigDict(
-    strict=True,
-    arbitrary_types_allowed=True,
-    extra="forbid",
-    validate_default=True,
-)
-"""Pydantic config for internal types: reject inputs that need conversion."""
-
-
-def validated_dataclass(
-    config: pydantic.ConfigDict,
-    make_error: Callable[[pydantic.ValidationError], Exception],
-) -> Callable[[type[DataclassT]], type[DataclassT]]:
-    """Create a class decorator that adds pydantic validation to a frozen dataclass.
-
-    The decorated class is re-created as a pydantic dataclass, so field values are
-    validated and converted according to their type annotations on every
-    instantiation (including via ``dataclasses.replace``). Annotations are resolved
-    at runtime, so this also works in modules using
-    ``from __future__ import annotations``.
-
-    Args:
-        config: The pydantic config that controls validation behavior.
-        make_error: Called with the raised ``pydantic.ValidationError`` to build the
-            exception that is raised in its place.
-
-    Returns:
-        A class decorator for frozen dataclasses.
-
-    """
-
-    def decorator(cls: type[DataclassT]) -> type[DataclassT]:
-        out = pydantic.dataclasses.dataclass(frozen=True, config=config)(cls)
-        # pydantic re-creates the class, which loses attributes that tooling and
-        # introspection rely on.
-        out.__doc__ = cls.__doc__
-        out.__annotations__ = dict(cls.__annotations__)
-        original_init = out.__init__
-
-        @functools.wraps(original_init)
-        def __init__(self: Any, *args: Any, **kwargs: Any) -> None:
-            try:
-                original_init(self, *args, **kwargs)
-            except pydantic.ValidationError as e:
-                raise make_error(e) from e
-
-        out.__init__ = __init__  # type: ignore[method-assign]
-        return typing.cast("type[DataclassT]", out)
-
-    return decorator
 
 
 def _algo_info_error(e: pydantic.ValidationError) -> Exception:
